+++ src/giskardpy/model/links.py
@@ -16,7 +16,8 @@ from giskardpy.model.utils import cube_volume, cube_surface, sphere_volume, cyli
 from giskardpy.my_types import PrefixName
 from giskardpy.my_types import my_string
 from giskardpy.utils.tfwrapper import np_to_pose
-from giskardpy.utils.utils import resolve_ros_iris, memoize
+from giskardpy.utils.utils import resolve_ros_iris
+from giskardpy.utils.decorators import memoize
 import giskardpy.casadi_wrapper as w
