+++ src/giskardpy/tree/behaviors/init_qp_controller.py
@@ -10,7 +10,7 @@ from giskardpy.goals.goal import Goal
 from giskardpy.qp.constraint import EqualityConstraint, InequalityConstraint, DerivativeInequalityConstraint
 from giskardpy.qp.qp_controller import QPProblemBuilder
 from giskardpy.tree.behaviors.plugin import GiskardBehavior
-from giskardpy.utils.utils import catch_and_raise_to_blackboard
+from giskardpy.utils.decorators import catch_and_raise_to_blackboard
 
 
 class InitQPController(GiskardBehavior):
