+++ src/giskardpy/tree/behaviors/send_trajectory_omni_drive_realtime.py
@@ -20,7 +20,7 @@ from giskardpy.qp.next_command import NextCommands
 from giskardpy.tree.behaviors.plugin import GiskardBehavior
 from giskardpy.utils import logging
 from giskardpy.utils.logging import loginfo
-from giskardpy.utils.utils import catch_and_raise_to_blackboard
+from giskardpy.utils.decorators import catch_and_raise_to_blackboard
 
 
 class SendTrajectoryToCmdVel(GiskardBehavior, ABC):
