+++ src/giskardpy/tree/behaviors/setup_base_traj_constraints.py
@@ -5,7 +5,7 @@ from py_trees import Status
 import giskardpy.identifier as identifier
 from giskardpy.goals.goal import Goal
 from giskardpy.tree.behaviors.plugin import GiskardBehavior
-from giskardpy.utils.utils import catch_and_raise_to_blackboard
+from giskardpy.utils.decorators import catch_and_raise_to_blackboard
 
 
 class SetDriveGoals(GiskardBehavior):
