+++ src/giskardpy/utils/decorators.py
@@ -1,0 +1,119 @@
+from __future__ import division
+
+# I only do this, because otherwise test/test_integration_pr2.py::TestWorldManipulation::test_unsupported_options
+# fails on github actions
+import urdf_parser_py.urdf as up
+
+import errno
+import inspect
+import json
+import os
+import pkgutil
+import sys
+import traceback
+from collections import OrderedDict
+from contextlib import contextmanager
+from copy import deepcopy
+from functools import wraps
+from time import time
+from typing import Type, Optional, Dict
+
+import numpy as np
+import roslaunch
+import rospkg
+import rospy
+import trimesh
+from genpy import Message
+from geometry_msgs.msg import PointStamped, Point, Vector3Stamped, Vector3, Pose, PoseStamped, QuaternionStamped, \
+    Quaternion
+from py_trees import Status, Blackboard
+from rospy_message_converter.message_converter import \
+    convert_ros_message_to_dictionary as original_convert_ros_message_to_dictionary, \
+    convert_dictionary_to_ros_message as original_convert_dictionary_to_ros_message
+from sensor_msgs.msg import JointState
+from visualization_msgs.msg import Marker, MarkerArray
+
+from giskardpy import identifier
+from giskardpy.exceptions import DontPrintStackTrace
+from giskardpy.god_map import GodMap
+from giskardpy.my_types import PrefixName
+from giskardpy.utils import logging
+from giskardpy.utils.time_collector import TimeCollector
+from giskardpy.utils.utils import has_blackboard_exception, raise_to_blackboard
+
+
+def memoize(function):
+    memo = function.memo = {}
+
+    @wraps(function)
+    def wrapper(*args, **kwargs):
+        # key = cPickle.dumps((args, kwargs))
+        # key = pickle.dumps((args, sorted(kwargs.items()), -1))
+        key = (args, frozenset(kwargs.items()))
+        try:
+            return memo[key]
+        except KeyError:
+            rv = function(*args, **kwargs)
+            memo[key] = rv
+            return rv
+
+    return wrapper
+
+
+def record_time(function):
+    return function
+    # god_map = GodMap()
+    time_collector: TimeCollector = god_map.get_data(identifier.timer_collector, default=TimeCollector())
+    if function.__name__ == 'solver_call':
+        @wraps(function)
+        def wrapper(self, *args, **kwargs):
+            qp_solver = self.solver_id
+            start_time = time()
+            result = function(self, *args, **kwargs)
+            time_delta = time() - start_time
+            time_collector.add_qp_solve_time(str(qp_solver), 0, 0, time_delta)
+            return result
+
+        return wrapper
+    else:
+        raise ValueError('Can only record time of \'solver_call\' functions.')
+
+
+def clear_memo(f):
+    if hasattr(f, 'memo'):
+        f.memo.clear()
+
+
+def copy_memoize(function):
+    memo = function.memo = {}
+
+    @wraps(function)
+    def wrapper(*args, **kwargs):
+        # key = cPickle.dumps((args, kwargs))
+        # key = pickle.dumps((args, sorted(kwargs.items()), -1))
+        key = (args, frozenset(kwargs.items()))
+        try:
+            return deepcopy(memo[key])
+        except KeyError:
+            rv = function(*args, **kwargs)
+            memo[key] = rv
+            return rv
+
+    return wrapper
+
+
+def catch_and_raise_to_blackboard(function):
+    @wraps(function)
+    def wrapper(*args, **kwargs):
+        if has_blackboard_exception():
+            return Status.FAILURE
+        try:
+            r = function(*args, **kwargs)
+        except Exception as e:
+            if not isinstance(e, DontPrintStackTrace):
+                traceback.print_exc()
+            raise_to_blackboard(e)
+            return Status.FAILURE
+        return r
+
+    return wrapper
